Feat: Item s3 이미지 업로드 구현을 위한 수정

+++ src/main/java/com/shoux_kream/item/dto/request/ItemSaveRequest.java
@@ -1,5 +1,6 @@
 package com.shoux_kream.item.dto.request;
 
+import com.shoux_kream.item.entity.Item;
 import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
@@ -13,7 +14,7 @@ public record ItemSaveRequest(
         String detailDescription,
         String imageKey,
         int inventory,
-        double price,
+        int price,
         List<String> searchKeywords, // 배열로 받기 위해 List<String> 사용
         Long itemInfoId, // ItemInfo의 ID만 받도록 수정
         int size
